requirements.txt
flask
watchdog
exifread
gunicorn
psutil
python-dotenv
flask-httpauth